# sgemm: simplify kernel_x86_avx logic and reduce shuffle overhead